+++ RELEASENOTES.md
@@ -1,15 +1,28 @@
 These are the release notes for the [mediawiki-datamodel](README.md).
 
-## Version 0.2 (under development)
+## Version 0.3 (under development)
 
-#### Compatability  changes
+#### Compatibility  changes
+
+* Revision objects now take a Content object as $content
+
+#### Additions
+
+* Content class
+* WikitextContent class
+
+#### Enhancements
+
+* Hashable and setable EditInfo class
+
+
+## Version 0.2 (2014-02-23)
+
+#### Compatibility  changes
 
 * Revision enhanced to allow more flexibility, Constructor and public functions have changed
 * contentmodel has been removed from the Page class
 
-#### Enhancments
-
-* Hashable and setable EditInfo class
 
 ## Version 0.1 (2014-02-23)